# swagger.yml
openapi: 3.0.0
info:
  title: Quinnn API List
  description: This is a reference api list of Quinnn Backend.
  version: 1.0.0
  contact:
    email: ankuragarwa246@gmail.com
servers:
  - url: https://lgbtq-backend-free.azurewebsites.net/api
    description: Production Server (Azure)
  - url: http://localhost:9999/api
    description: Development server
paths:
...
  /user/profile:
    get:
      tags:
        - User
      summary: Get Profile for the logged in user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /user/profile/update:  
    put:
      tags:
        - User
      summary: Update Profile
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/UserProfile'
      responses:
        '200':
          description: Profile Updated Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /user/search/{username}:
    get:
      tags:
        - User
      summary: Search Profile for a particular username
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: username
          description: Username of the user
          schema:
            type: string
      responses:
        '200':
          description: Profile Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /user/follow/{userId}:
    put:
      tags:
        - User
      summary: Follow a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          description: User ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: User Followed Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /user/unfollow/{userId}:
    put:
      tags:
        - User
      summary: Unfollow a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          description: User ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: User Unfollowed Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        bio:
          type: string
        profileImageURL:
          type: string
        coverImageURL:
          type: string
        dob:
          type: string
          format: date
        country:
          type: string
        isPrivate:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
...
    UserProfile:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateProfileInput'
        description: Update Profile
...